Refactor BigQuery connection on read side
